+++ pkg/dtd_impl/lib/src/dtd_stream_manager.dart
@@ -10,6 +10,20 @@ import '../dart_tooling_daemon.dart';
 class DTDStreamManager extends StreamManager {
   DTDStreamManager(this.dtd);
 
+  /// The name of the stream for events related to new services/methods being
+  /// added and removed.
+  ///
+  /// This stream is not part of the VM service protocol.
+  static const servicesStreamId = 'Service';
+
+  /// The kind of the event sent over the [servicesStreamId] stream when a new
+  /// service method is registered.
+  static const serviceRegisteredId = 'ServiceRegistered';
+
+  /// The kind of the event sent over the [servicesStreamId] stream when a new
+  /// service method is unregistered.
+  static const serviceUnregisteredId = 'ServiceUnregistered';
+
   final DartToolingDaemon dtd;
 
   /// Send an event to the [stream].
@@ -29,7 +43,7 @@ class DTDStreamManager extends StreamManager {
   ) {
     super.postEvent(
       stream,
-      <String, dynamic>{
+      <String, Object?>{
         'streamId': stream,
         'eventKind': eventKind,
         'eventData': eventData,